fixed code of DIalog component
