+++ src/components/Dialog/ModalFooter.Component.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { ModalFooterProperties } from '@blueeast/bluerain-ui-interfaces';
+import {  ModalFooter } from 'reactstrap';
+
+const  ModalFooterComponent =(props:ModalFooterProperties)=> {
+    return(
+        <ModalFooter>
+            {props.children}
+        </ModalFooter>
+    );
+}
+export default ModalFooterComponent;
