+++ src/components/Dialog/Test/Modal.stories.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react';
+import DialogComponent from '../ModalStory';
+import { storiesOf } from '@storybook/react';
+const stories = storiesOf('Modal', module);
+
+stories.add('ModalComponent ', () => {
+	return <DialogComponent   />;
+});
